+++ irs-registry-client/src/test/java/org/eclipse/tractusx/irs/registryclient/central/CentralDigitalTwinRegistryServiceTest.java
@@ -37,6 +37,7 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
+import org.eclipse.tractusx.irs.SemanticModelNames;
 import org.eclipse.tractusx.irs.component.Shell;
 import org.eclipse.tractusx.irs.component.assetadministrationshell.AssetAdministrationShellDescriptor;
 import org.eclipse.tractusx.irs.component.assetadministrationshell.Endpoint;
@@ -57,8 +58,6 @@ import org.springframework.web.client.RestClientException;
 @ExtendWith(MockitoExtension.class)
 class CentralDigitalTwinRegistryServiceTest extends LocalTestDataConfigurationAware {
 
-    private final String singleLevelBomAsBuiltURN = "urn:samm:io.catenax.single_level_bom_as_built:3.0.0#SingleLevelBomAsBuilt";
-    private final String serialPartURN = "urn:samm:io.catenax.serial_part:3.0.0#SerialPart";
     private DigitalTwinRegistryService digitalTwinRegistryService;
     @Mock
     private DigitalTwinRegistryClient dtRegistryClientMock;
@@ -92,8 +91,9 @@ class CentralDigitalTwinRegistryServiceTest extends LocalTestDataConfigurationAw
 
         assertThat(endpoint.getProtocolInformation().getSubprotocolBody()).contains(existingCatenaXId);
         assertThat(shellEndpoints.get(0).getSemanticId().getKeys().get(0).getValue()).isEqualTo(
-                singleLevelBomAsBuiltURN);
-        assertThat(shellEndpoints.get(1).getSemanticId().getKeys().get(0).getValue()).isEqualTo(serialPartURN);
+                SemanticModelNames.SINGLE_LEVEL_BOM_AS_BUILT_3_0_0);
+        assertThat(shellEndpoints.get(1).getSemanticId().getKeys().get(0).getValue()).isEqualTo(
+                SemanticModelNames.SERIAL_PART_3_0_0);
     }
 
     @Test
@@ -187,6 +187,7 @@ class CentralDigitalTwinRegistryServiceTest extends LocalTestDataConfigurationAw
         assertThat(shellEndpoints).isNotNull().isNotEmpty();
         final SubmodelDescriptor endpoint = shellEndpoints.get(0);
 
-        assertThat(endpoint.getSemanticId().getKeys().get(0).getValue()).isEqualTo(singleLevelBomAsBuiltURN);
+        assertThat(endpoint.getSemanticId().getKeys().get(0).getValue()).isEqualTo(
+                SemanticModelNames.SINGLE_LEVEL_BOM_AS_BUILT_3_0_0);
     }
 }